# docs/swagger.yaml
basePath: /v1
definitions:
  github_com_pickle_pw_monolith_db_sqlc.ContentCategory:
    enum:
    - games
    - movies
    - video
    - anime
    - series
    - custom
    - any
    type: string
    x-enum-varnames:
    - ContentCategoryGames
    - ContentCategoryMovies
    - ContentCategoryVideo
    - ContentCategoryAnime
    - ContentCategorySeries
    - ContentCategoryCustom
    - ContentCategoryAny
...
  github_com_pickle_pw_monolith_internal_types.ContentRes:
    properties:
      category:
        $ref: '#/definitions/github_com_pickle_pw_monolith_db_sqlc.ContentCategory'
      id:
        type: string
      title:
        type: string
      userId:
        type: string
    type: object
...
host: api.pickle.pw
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Pickle server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pickle API
  version: "1.0"
paths:
...
  /v1/content/{category}:
    get:
      consumes:
      - application/json
      description: Search content
      parameters:
      - description: Query
        in: query
        name: query
        required: true
        type: string
      - description: User ID
        in: query
        name: userId
        type: string
      - description: Category
        in: path
        name: category
        required: true
        type: string
      - description: Locale
        in: query
        name: locale
        type: string
      - description: Page
        in: query
        name: page
        type: string
      - description: Size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.ContentRes'
            type: array
      summary: Search content
      tags:
      - content
  /v1/content/{category}/{id}:
    get:
      consumes:
      - application/json
      description: Get content by ID
      parameters:
      - description: Content ID
        in: path
        name: id
        required: true
        type: string
      - description: Category
        in: path
        name: category
        required: true
        type: string
      - description: Locale
        in: query
        name: locale
        type: string
      - description: Cover Size
        in: query
        name: coverSize
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get content by ID
      tags:
      - content
...
  /v1/users/{userId}/content/search:
    get:
      consumes:
      - application/json
      description: Search content
      parameters:
      - description: Query
        in: query
        name: query
        required: true
        type: string
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.ContentRes'
            type: array
      summary: Search content
      tags:
      - content
...
swagger: "2.0"